requirements.txt
fastapi~=0.95.2
starlette~=0.27.0
h11
httplib2
idna
passlib~=1.7.4
pip
protobuf==4.22.1
requests
requests-oauthlib
setuptools==67.6.0
six
sniffio
starlette
psycopg2~=2.9.6
PyYAML~=6.0
anyio
bcrypt
cachetools
SQLAlchemy~=2.0.7
certifi
charset-normalizer
click
colorama
ecdsa
transliterate
typing_extensions
uritemplate
urllib3
uvicorn~=0.22.0
wheel==0.40.0
pydantic~=1.10.7
fastapi-jwt-auth
PyJWT~=2.7.0